reservation activity, repository and logic

+++ app/src/main/java/com/example/macc_project_app/data/reservation/Reservation.kt
@@ -3,5 +3,5 @@ package com.example.macc_project_app.data.reservation
 data class Reservation(
     val id: Long,
     val restaurantId: Long,
-    val numberSeats: Long
+    val numberSeats: Int
 )
